Extract ALSurfaceContext so it can be reused in other places.
We are going to need to expose current Surface value in various
applications.
Since in this repo we cannot rely on any instance of React, ReactDOM,
... we have to wait for an init to first pass these modules and then
have the right functionality setup.
